// app/src/api/models/ExperimentIntersectionPairCountsExperiments.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Snowman API
 * _This document describes the REST API of the snowman data matching benchmark tool._  Comparing data matching algorithms is still an unsolved topic in both industry and research.  With snowman, developers and researchers will be able to compare the performance of different data matching  solutions or improve new algorithms. 
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: snowman@groups.sap.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ExperimentIntersectionPairCountsExperiments
 */
export interface ExperimentIntersectionPairCountsExperiments {
    /**
     * 
     * @type {number}
     * @memberof ExperimentIntersectionPairCountsExperiments
     */
    experimentId: number;
    /**
     * 
     * @type {boolean}
     * @memberof ExperimentIntersectionPairCountsExperiments
     */
    predictedCondition: boolean;
}

export function ExperimentIntersectionPairCountsExperimentsFromJSON(json: any): ExperimentIntersectionPairCountsExperiments {
    return ExperimentIntersectionPairCountsExperimentsFromJSONTyped(json, false);
}

export function ExperimentIntersectionPairCountsExperimentsFromJSONTyped(json: any, ignoreDiscriminator: boolean): ExperimentIntersectionPairCountsExperiments {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'experimentId': json['experimentId'],
        'predictedCondition': json['predictedCondition'],
    };
}

export function ExperimentIntersectionPairCountsExperimentsToJSON(value?: ExperimentIntersectionPairCountsExperiments | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'experimentId': value.experimentId,
        'predictedCondition': value.predictedCondition,
    };
}